separate declaration of lexical_token_t
